requirements.txt
   streamlit
   pandas
   numpy
   matplotlib
   seaborn
   plotly
   PyGithub
   python-dotenv
   requests
   tqdm
   pycountry
   geopy